requirements.txt
python-telegram-bot>=20.8
aiosqlite>=0.19.0
python-dotenv>=1.0.0
aiocryptopay>=0.3.3
